separate iterator and count
